services/policy-engine/requirements.txt
zen-engine==0.53.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
kafka-python==2.0.2
python-dotenv==1.0.0
requests==2.31.0
